refactor: split dashboard overview and smart-match shell
Extract overview columns, filter/sort, data table, and bulk-import helpers; split Smart Match into filters, search, results, and helpers so both shells stay compositional.
